requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
requests>=2.31.0
pyyaml>=6.0.1
aiofiles>=23.2.1
tqdm>=4.66.0
psutil>=5.9.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0

# Quality Control
detoxify>=0.5.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
datasketch>=1.6.0

# Analytics
pandas>=2.1.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Export Formats
huggingface_hub>=0.20.0
datasets>=2.16.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.0

# MLOps (optional)
celery>=5.3.0
redis>=5.0.0
mlflow>=2.10.0
wandb>=0.16.0
apscheduler>=3.10.0